MyService Beta Release
The Menu have been completed and connected with the rest of the code.

Updated Payment Control to get builtin user id.

Wallet section of the payment component have been updated adding options to view the balance.

Refund section of the Transaction component updated to hold the user id and the refund status and allows the user to view his refund requests

Updates to the login and sign up form

updates to the user control to add more features for the current user